move ball and paddle to their own modules
