+++ pettingzoo/butterfly/cooperative_pong/cooperative_pong.py
@@ -10,8 +10,10 @@ from pettingzoo.utils import wrappers
 from pettingzoo.utils.agent_selector import agent_selector
 from pettingzoo.utils.conversions import parallel_wrapper_fn
 
+from .ball import Ball
 from .cake_paddle import RENDER_RATIO, CakePaddle
 from .manual_control import manual_control
+from .paddle import Paddle
 
 KERNEL_WINDOW_LENGTH = 2
 
@@ -50,135 +52,6 @@ def get_valid_angle(randomizer):
     return angle
 
 
-def get_small_random_value(randomizer):
-    # generates a small random value between [0, 1/100)
-    return (1 / 100) * randomizer.rand()
-
-
-class PaddleSprite(pygame.sprite.Sprite):
-    def __init__(self, dims, speed):
-        self.surf = pygame.Surface(dims)
-        self.rect = self.surf.get_rect()
-        self.speed = speed
-
-    def reset(self):
-        pass
-
-    def draw(self, screen):
-        pygame.draw.rect(screen, (255, 255, 255), self.rect)
-
-    def update(self, area, action):
-        # action: 1 - up, 2 - down
-        movepos = [0, 0]
-        if action > 0:
-            if action == 1:
-                movepos[1] = movepos[1] - self.speed
-            elif action == 2:
-                movepos[1] = movepos[1] + self.speed
-
-            # make sure the players stay inside the screen
-            newpos = self.rect.move(movepos)
-            if area.contains(newpos):
-                self.rect = newpos
-
-    def process_collision(self, b_rect, dx, dy, b_speed, paddle_type):
-        '''
-
-        Parameters
-        ----------
-        b_rect : Ball rect
-        dx, dy : Ball speed along single axis
-        b_speed : Ball speed
-
-        Returns
-        -------
-        is_collision: 1 if ball collides with paddle
-        b_rect: new ball rect
-        b_speed: new ball speed
-
-        '''
-        if not self.rect.colliderect(b_rect):
-            return False, b_rect, b_speed
-        # handle collision from left or right
-        if paddle_type == 1 and b_rect.left < self.rect.right:
-            b_rect.left = self.rect.right
-            if b_speed[0] < 0:
-                b_speed[0] *= -1
-        elif paddle_type == 2 and b_rect.right > self.rect.left:
-            b_rect.right = self.rect.left
-            if b_speed[0] > 0:
-                b_speed[0] *= -1
-        # handle collision from top
-        if b_rect.bottom > self.rect.top and b_rect.top - dy < self.rect.top and b_speed[1] > 0:
-            b_rect.bottom = self.rect.top
-            if b_speed[1] > 0:
-                b_speed[1] *= -1
-        # handle collision from bottom
-        elif b_rect.top < self.rect.bottom and b_rect.bottom - dy > self.rect.bottom and b_speed[1] < 0:
-            b_rect.top = self.rect.bottom - 1
-            if b_speed[1] < 0:
-                b_speed[1] *= -1
-        return True, b_rect, b_speed
-
-
-class BallSprite(pygame.sprite.Sprite):
-    def __init__(self, randomizer, dims, speed, bounce_randomness=False):
-        self.surf = pygame.Surface(dims)
-        self.rect = self.surf.get_rect()
-        self.speed_val = speed
-        self.speed = [int(self.speed_val * np.cos(np.pi / 4)), int(self.speed_val * np.sin(np.pi / 4))]
-        self.bounce_randomness = bounce_randomness
-        self.done = False
-        self.hit = False
-        self.randomizer = randomizer
-
-    def update2(self, area, p0, p1):
-        return self.move(self.speed[0], self.speed[1], area, p0, p1)
-
-    def move(self, dx, dy, area, p0, p1):
-        # move ball rect
-        self.rect.x += dx
-        self.rect.y += dy
-
-        if not area.contains(self.rect):
-            # bottom wall
-            if self.rect.bottom > area.bottom:
-                self.rect.bottom = area.bottom
-                self.speed[1] = -self.speed[1]
-            # top wall
-            elif self.rect.top < area.top:
-                self.rect.top = area.top
-                self.speed[1] = -self.speed[1]
-            # right or left walls
-            else:
-                return True
-                self.speed[0] = -self.speed[0]
-
-        else:
-            # Do ball and bat collide?
-            # add some randomness
-            r_val = 0
-            if self.bounce_randomness:
-                r_val = get_small_random_value(self.randomizer)
-
-            # ball in left half of screen
-            if self.rect.center[0] < area.center[0]:
-                is_collision, self.rect, self.speed = p0.process_collision(self.rect, dx, dy, self.speed, 1)
-                if is_collision:
-                    self.speed = [self.speed[0] + np.sign(self.speed[0]) * r_val, self.speed[1] + np.sign(self.speed[1]) * r_val]
-            # ball in right half
-            else:
-                is_collision, self.rect, self.speed = p1.process_collision(self.rect, dx, dy, self.speed, 2)
-                if is_collision:
-                    self.speed = [self.speed[0] + np.sign(self.speed[0]) * r_val, self.speed[1] + np.sign(self.speed[1]) * r_val]
-
-        return False
-
-    def draw(self, screen):
-        # screen.blit(self.surf, self.rect)
-        pygame.draw.rect(screen, (255, 255, 255), self.rect)
-
-
 class CooperativePong:
     def __init__(self, randomizer, ball_speed=9, left_paddle_speed=12, right_paddle_speed=12, cake_paddle=True, max_cycles=900, bounce_randomness=False, max_reward=100, off_screen_penalty=-10):
         super().__init__()
@@ -209,16 +82,16 @@ class CooperativePong:
         self.max_cycles = max_cycles
 
         # paddles
-        self.p0 = PaddleSprite((20 // RENDER_RATIO, 80 // RENDER_RATIO), left_paddle_speed)
+        self.p0 = Paddle((20 // RENDER_RATIO, 80 // RENDER_RATIO), left_paddle_speed)
         if cake_paddle:
             self.p1 = CakePaddle(right_paddle_speed)
         else:
-            self.p1 = PaddleSprite((20 // RENDER_RATIO, 100 // RENDER_RATIO), right_paddle_speed)
+            self.p1 = Paddle((20 // RENDER_RATIO, 100 // RENDER_RATIO), right_paddle_speed)
 
         self.agents = ["paddle_0", "paddle_1"]  # list(range(self.num_agents))
 
         # ball
-        self.ball = BallSprite(randomizer, (20 // RENDER_RATIO, 20 // RENDER_RATIO), ball_speed, bounce_randomness)
+        self.ball = Ball(randomizer, (20 // RENDER_RATIO, 20 // RENDER_RATIO), ball_speed, bounce_randomness)
         self.randomizer = randomizer
 
         self.reinit()
